implement Makefile for the project

+++ CgiDirective.cpp
@@ -1,20 +1,18 @@
 #include "CgiDirective.hpp"
 
-CgiDirective::CgiDirective()
-{
-}
+CgiDirective::CgiDirective() { }
 
 void	CgiDirective::setExtensions(const std::vector<std::string> &content)
 {
     if (content.empty())
         return;
-    for (int i = 0; i < content.size();i++)
+    for (size_t i = 0; i < content.size();i++)
     {
         if (content[i].find('.') == std::string::npos || content[i].size() <= 2 || content[i].substr(0, 2) != "*.") // optional give a max length 
             throw (std::runtime_error("invalid cgi extension in \"cgi_extension\" directive"));
     }
     this->_isEnabled = true;
-    for (int i = 0; i < content.size(); i++)
+    for (size_t i = 0; i < content.size(); i++)
         this->extensions.push_back(content[i].substr(1, content[i].length()));
 }
 
@@ -28,7 +26,4 @@ bool CgiDirective::isEnabled()
     return (this->_isEnabled);
 }
 
-CgiDirective::~CgiDirective()
-{
-
-}
+CgiDirective::~CgiDirective() { }
